Generate acceptance and integration tests for properties
