+++ tests/dummy/app/templates/components/clickable-component.hbs
@@ -1,0 +1,1 @@
+{{{html}}}
