+++ tests/dummy/app/templates/components/html-render.hbs
@@ -1,0 +1,1 @@
+{{{html}}}
